feat: do not show animation on balance for low end devices
